+++ changelog.md
@@ -3,6 +3,7 @@
   - Forces wedding veils to use their 'inside sanctum of the twelve' model.
 - New Tweak: `Extended Desynthesis Window`
   - Shows the current and optimal desynthesis level on the `/ac desynthesis` window.
+- `UI -> Large Cooldown Counter` Add support for Cross Hotbar
 
 ## 1.3.3.0
 - `UI -> Custom Time Format` Allow for custom order of clocks.